Split out functionality behind ModelTransformer into its ownWings::ORMConverter class.
