+++ lib/wings.rb
@@ -21,6 +21,7 @@ module Wings; end
 
 require 'valkyrie'
 require 'wings/model_transformer'
+require 'wings/orm_converter'
 require 'wings/valkyrizable'
 require 'wings/valkyrie/metadata_adapter'
 require 'wings/valkyrie/resource_factory'
@@ -31,7 +32,7 @@ require 'wings/valkyrie/storage/active_fedora'
 ActiveFedora::Base.include Wings::Valkyrizable
 
 Valkyrie.config.resource_class_resolver = lambda do |_klass_name|
-  Wings::ModelTransformer.convert_class_name_to_valkyrie_resource_class(internal_resource)
+  Wings::OrmConverter.convert_class_name_to_valkyrie_resource_class(internal_resource)
 end
 
 Valkyrie::MetadataAdapter.register(
