+++ lib/wings/transformer_value_mapper.rb
@@ -21,8 +21,8 @@ module Wings
     def result
       attributes = value.attributes.symbolize_keys
       nested_object = Wings::ActiveFedoraConverter::NestedResource.new(attributes)
-      klass = Wings::ModelTransformer::ResourceClassCache.instance.fetch(Wings::ActiveFedoraConverter::NestedResource) do
-        ModelTransformer.to_valkyrie_resource_class(klass: nested_object.class)
+      klass = Wings::OrmConverter::ResourceClassCache.instance.fetch(Wings::ActiveFedoraConverter::NestedResource) do
+        OrmConverter.to_valkyrie_resource_class(klass: nested_object.class)
       end
       klass.new(attributes)
     end
